foodai/pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "azumio-foodai"
dependencies = [
  'requests',
  'pillow',
  'dacite',
]
version = "1.0.0"
authors = [
  { name="Igor Rendulic", email="igor@azumio.com" },
]
description = ""
readme = "README.md"
license = { file="LICENSE" }
requires-python = ">=3.7"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: Apache Software License",
    "Operating System :: OS Independent",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Intended Audience :: Developers",
    "Natural Language :: English",
]

[project.urls]
"Homepage" = "https://github.com/azumio/foodai"
"Bug Tracker" = "https://github.com/azumio/foodai/issues"